client/pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "streamvis"
version = "0.1.0"
description = "Interactive visualizations of streaming data using Bokeh"
authors = [{name="Henry Bigelow", email="hrbigelow@gmail.com"}]
requires-python = ">=3.12.0"
dependencies = [
  "tornado",
  "numpy",
  "bokeh==3.7.1",
  "google-cloud-storage",
  "hydra-core>=1.3.2",
  "grpcio-tools==1.73.1",
  "vulture>=2.14",
  "fire>=0.7.1",
  "pydantic>=2.12.5",
]

[project.scripts]
streamvis = "streamvis.script:main"

[tool.setuptools]
packages = ["streamvis"]



